addedd validations to CreateUpdateRequests,migrations for alert_notifications

+++ construction/views.py
@@ -626,16 +626,25 @@ class CreateUpdateDrawRequest(APIView):
             else:
                 val = draw_req_obj.order_by('-draw_request').values_list('draw_request',flat=True)
                 draw_request = val[0] +1
-                print("budget_master_obj",budget_master_obj)
+                # Check the status of the previous draw from the DrawTracking table
+                last_draw = DrawTracking.objects.get(loan_id=loan_id,draw_request=val[0])
+                # If the last draw exists and its status is "Pending" or "In Review", prevent new draw creation
+                if last_draw and last_draw.draw_status in ["Pending", "In Review"]:
+                    return Response(
+                         {"error":"The previous draw is still 'Pending' or 'In Review'. Cannot create a new draw."},
+                         status=status.HTTP_400_BAD_REQUEST
+                    )
+
+                
                 for obj in budget_master_obj:
-                    print("budget amount",obj[1])
                     budget_amount = obj[1]  
                     released_amount_previous = DrawRequest.objects.filter(budget_master_id=obj).values_list('released_amount',flat=True)        
-                    print(released_amount_previous)
                     released_amount_list = list(released_amount_previous)
-                    print(released_amount_list)
-
-                    released_amount=released_amount_list[0]
+                    if released_amount_list:
+                        released_amount=released_amount_list[0]
+                    else:
+                        released_amount = 0
+                    
                     new_instance = DrawRequest(
                         budget_master_id=obj[0],
                         draw_request = draw_request,                    
@@ -766,12 +775,15 @@ class RetrieveDeleteUpdateDrawTracking(APIView):
     def delete(self, request, id):
         try:
             draw_tracking_obj = DrawTracking.objects.get(pk=id)
-            draw_request = draw_tracking_obj.draw_request
-            loan_id = draw_tracking_obj.loan
-            draw_tracking_obj.delete()
-            budget_master_ids = BudgetMaster.objects.filter(loan_id=loan_id).values_list('id',flat=True)
-            DrawRequest.objects.filter(draw_request=draw_request, budget_master_id__in = budget_master_ids).delete()
-            return Response(status=status.HTTP_204_NO_CONTENT)
+            if draw_tracking_obj.draw_status in ['Pending','In Review']: 
+                draw_request = draw_tracking_obj.draw_request
+                loan_id = draw_tracking_obj.loan
+                draw_tracking_obj.delete()
+                budget_master_ids = BudgetMaster.objects.filter(loan_id=loan_id).values_list('id',flat=True)
+                DrawRequest.objects.filter(draw_request=draw_request, budget_master_id__in = budget_master_ids).delete()
+                return Response(status=status.HTTP_204_NO_CONTENT)
+            else:
+                return Response({"error":"Only draw requests with 'Pending' or 'In Review' status can be deleted."},status=status.HTTP_403_FORBIDDEN)
         except DrawTracking.DoesNotExist:
             return Response({"error": "DrawTracking object not found."}, status=status.HTTP_404_NOT_FOUND)
         except Exception as e:
